+++ src/components/WizardConfirmationModal.js
@@ -6,10 +6,14 @@ import BraintreeCreditCardForm from './BraintreeCreditCardForm';
 export default class WizardConfirmationModal extends Component {
     render() {
         const {
+            startDate,
+
             getToken,
             onSubmit,
             handleClose
         } = this.props;
+
+        const start = (startDate && startDate.format('MM/DD/YYYY')) || '...';
 
         return (<div className="modal payment-modal fade in show" id="pmtModal" role="dialog">
             <div className="modal-dialog" role="document">
@@ -21,13 +25,17 @@ export default class WizardConfirmationModal extends Component {
                             aria-label="Close">
                             <span aria-hidden="true">×</span>
                         </button>
-                        <h1 className="modal-title" id="myPaymentModal">Enter Payment Details</h1>
-                        <h4>to launch and reach 2,000 users</h4>
+                        <h1 className="modal-title" id="myPaymentModal">Start Your FREE Trial</h1>
+                        <h4>You will not be charged until after your free trial ends on {start}.
+                            Cancel anytime.</h4>
                     </div>
                     <div className="modal-body text-center">
-                        <BraintreeCreditCardForm submitText="Start Promoting My App"
+                        <BraintreeCreditCardForm submitText="Start Your Free Trial"
                             getToken={getToken}
                             onSubmit={onSubmit} />
+                        <div className="payment-charge-info">
+                            Reach 2,000 users for only $49.99/month after your trial
+                        </div>
                         <div className="secured-locked">
                             <div className="signup__secured">
                                 <span className="signup__secured--icon">
@@ -41,7 +49,7 @@ export default class WizardConfirmationModal extends Component {
                         </div>
                     </div>
                     <div className="modal-footer">
-                        <p>Your information is 128-bit SSL encrypted</p>
+                        <p>Your information is secured with 256-bit RSA encryption</p>
                     </div>
                 </div>
             </div>
@@ -52,5 +60,8 @@ export default class WizardConfirmationModal extends Component {
 WizardConfirmationModal.propTypes = {
     getToken: PropTypes.func.isRequired,
     onSubmit: PropTypes.func.isRequired,
-    handleClose: PropTypes.func.isRequired
+    handleClose: PropTypes.func.isRequired,
+    startDate: PropTypes.shape({
+        format: PropTypes.func.isRequired
+    })
 };